pyproject.toml
[tool.poetry]
name = "OrderServiceAPI"
version = "0.1.0"
description = ""
authors = ["'Kamran <kamrankek48@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.112.0"}
dependency-injector = "^4.41.0"
uvicorn = "^0.30.6"
sqlalchemy = {extras = ["acyncio"], version = "^2.0.32"}
aiosqlite = "^0.20.0"
alembic = "^1.13.2"
aiohttp = "^3.10.9"
strawberry-graphql = {extras = ["fastapi"], version = "^0.239.2"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
requests = "^2.32.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"